src/requirements.txt
nonebot2>=2.0.0
nonebot-adapter-onebot11>=2.0.0
nonebot-plugin-datastore>=1.0.0
httpx>=0.24.0
qrcode>=7.0
pillow>=10.0.0
jinja2>=3.0.0
fastapi>=0.68.0
uvicorn>=0.15.0